ajustes

+++ SolutionName/SolutionName.Api/Controllers/ClienteController.cs
@@ -38,7 +38,7 @@ public class ClienteController : ControllerBase
     [Route("/{id}/cotacao")]
     public async Task<IActionResult> Patch(Guid id,[FromBody] PatchCotacaoRequest request)
     {
-        var result = await _ClienteService.Patch(id,request.valorCotadoEmReais);
+        var result = await _ClienteService.Patch(id,(decimal)request.valorCotadoEmReais);
 
         var response = new PatchCotacaoResponse(
             new PatchClienteResponse(result.nome, result.email, result.id),
